refactor: refactoring the giftsDisplay structure layout and wedding header
