+++ frontend/src/app/giftList/[id]/(components)/conditionalRender/conditionalRender.tsx
@@ -1,8 +1,8 @@
 'use client'
 
-import GuestList from '@/app/giftList/[id]/(components)/giftsListDisplay/GuestList'
-import GuestRequest from '@/app/giftList/[id]/(components)/giftsListDisplay/GuestRequest'
-import OwnerList from '@/app/giftList/[id]/(components)/giftsListDisplay/OwnerList'
+import GuestList from '@/app/giftList/[id]/(components)/giftsListDisplay/guest-list'
+import GuestRequest from '@/app/giftList/[id]/(components)/giftsListDisplay/guest-request'
+import OwnerList from '@/app/giftList/[id]/(components)/giftsListDisplay/owner-list'
 import { useContextWrap } from '@/contextAPI/context'
 import useGetData from '@/app/giftList/[id]/(hooks)/useGetData'
 
@@ -21,7 +21,7 @@ export default function ConditionalRenderingListPage() {
         )}
         {isCreator && (
           <>
-            <div className="flex flex-col items-center">
+            <div className="flex flex-col">
               <OwnerList />
             </div>
           </>
